LLM バックエンドと制御クラスのモジュール分割
