+++ backend/ai_meeting.py
@@ -11,282 +11,17 @@ from typing import Dict, List, Optional, Tuple
 import math
 import traceback
 import random
-from pydantic import BaseModel
 from pathlib import Path
 from datetime import datetime
 
 from backend.ai_meeting.config import AgentConfig, MeetingConfig, Turn
+from backend.ai_meeting.controllers import KPIFeedback, Monitor, PendingTracker, ShockEngine
 from backend.ai_meeting.evaluation import KPIEvaluator
+from backend.ai_meeting.llm import LLMRequest, OllamaBackend, OpenAIBackend
 from backend.ai_meeting.logging import LiveLogWriter
 from backend.ai_meeting.metrics import MetricsLogger
 from backend.ai_meeting.utils import banner, clamp
 
-# ===== LLM Backends =====
-
-class LLMRequest(BaseModel):
-    system: str
-    messages: List[Dict[str, str]]
-    temperature: float = 0.7
-    max_tokens: int = 800
-
-class LLMBackend:
-    def generate(self, req: LLMRequest) -> str:
-        raise NotImplementedError
-
-# --- OpenAI backend ---
-class OpenAIBackend(LLMBackend):
-    def __init__(self, model: Optional[str] = None):
-        try:
-            from openai import OpenAI
-            self.client = OpenAI()
-        except ImportError as e:
-            raise RuntimeError("OpenAI backend requires 'openai' package. Please `pip install openai` or use `--backend ollama`.") from e
-        self.model = model or os.getenv("OPENAI_MODEL", "gpt-4o-mini")
-
-    def generate(self, req: LLMRequest) -> str:
-        # OpenAI Chat Completions
-        messages: list[dict[str, str]] = [{"role":"system","content":req.system}] + req.messages
-        
-        resp = self.client.chat.completions.create(
-            model=self.model,
-            messages=typing.cast(typing.Iterable[typing.Any], messages),
-            temperature=req.temperature,
-            max_tokens=req.max_tokens,
-        )
-        return (resp.choices[0].message.content or "").strip()
-
-# --- Ollama backend (local) ---
-class OllamaBackend(LLMBackend):
-    def __init__(self, model: str = "llama3", host: str = "http://localhost:11434"):
-        import requests
-        self.requests = requests
-        self.model = model
-        self.host = host
-        if not self.host.startswith("http://localhost"):
-            raise RuntimeError("Ollama host must be localhost for 100% local run.")
-
-    def generate(self, req: LLMRequest) -> str:
-        url = f"{self.host}/api/chat"
-        payload = {
-            "model": self.model,
-            "messages": [{"role":"system","content":req.system}] + req.messages,
-            "options": {"temperature": req.temperature},
-            "stream": False,
-        }
-        r = self.requests.post(url, json=payload, timeout=600)
-        r.raise_for_status()
-        data = r.json()
-        return data.get("message",{}).get("content","").strip()
-
-
-# ---- Step 4: Monitor（裏方AI：フェーズ検知） ----
-class Monitor:
-    """
-    ルールベースMVP：
-      - 直近W発言の“まとまり度（cohesion）”をトークンJaccardの平均で近似
-      - 未解決件数が W 内で一定割合以上 減っている
-      - もしくは高類似（>=0.9相当）のループが K 回連続
-    満たしたらフェーズ確定。本文には何も挿入せず、phases.jsonl にだけ記録。
-    """
-    def __init__(self, cfg: MeetingConfig):
-        self.cfg = cfg
-        self._last_turn_idx = 0
-        self._loop_streak = 0
-
-    def observe(self, history: List[Turn], unresolved_hist: List[int], window: int) -> Optional[Dict]:
-        if len(history) - self._last_turn_idx < 1:
-            return None
-        self._last_turn_idx = len(history)
-        W = min(window, len(history))
-        if W < 3:
-            return None
-        recent = history[-W:]
-        # まとまり度（Jaccard平均）
-        sets = [self._token_set(t.content) for t in recent]
-        sim_sum = 0.0; cnt = 0
-        for i in range(W-1):
-            for j in range(i+1, W):
-                sim_sum += self._jacc(sets[i], sets[j]); cnt += 1
-        cohesion = (sim_sum/cnt) if cnt else 0.0
-        # ループ監視（直近2発言の高類似）
-        loop_hit = 0.0
-        if len(recent) >= 2:
-            sA, sB = sets[-1], sets[-2]
-            loop_hit = self._jacc(sA, sB)
-            self._loop_streak = self._loop_streak + 1 if loop_hit >= 0.90 else 0
-        # 未解決の減少率（正なら減っている）
-        unresolved_drop = 0.0
-        if len(unresolved_hist) >= 2:
-            first = unresolved_hist[0]; last = unresolved_hist[-1]
-            if first > 0:
-                unresolved_drop = max(0.0, (first - last) / first)
-        # 判定
-        reason = None
-        if self._loop_streak >= self.cfg.phase_loop_threshold:
-            reason = "loop"
-        elif cohesion >= self.cfg.phase_cohesion_min and unresolved_drop >= self.cfg.phase_unresolved_drop:
-            reason = "cohesion_unresolved"
-        if not reason:
-            return None
-        # 要約（裏方。本文に挿入しない）
-        summary = self._summarize_phase(recent)
-        return {
-            "start_turn": len(history)-W+1,
-            "end_turn": len(history),
-            "cohesion": round(cohesion,3),
-            "unresolved_drop": round(unresolved_drop,3),
-            "loop_streak": self._loop_streak,
-            "reason": reason,
-            "summary": summary,
-        }
-
-    def _summarize_phase(self, turns: List[Turn]) -> str:
-        # LLMを使わず簡潔な要約（先頭/末尾/頻出語を抽出）。将来はLLM要約に差し替え可。
-        texts = [t.content for t in turns]
-        head = texts[0][:60]; tail = texts[-1][:60]
-        return f"フェーズ要約: 先頭『{head}…』→末尾『{tail}…』"
-
-    @staticmethod
-    def _token_set(text: str) -> set:
-        t = re.sub(r"[0-9]+", " ", text)
-        t = re.sub(r"[^\w\u3040-\u30ff\u4e00-\u9fff]+", " ", t, flags=re.UNICODE)
-        toks = [w for w in t.lower().split() if len(w) > 1]
-        return set(toks)
-
-    @staticmethod
-    def _jacc(a:set, b:set) -> float:
-        if not a or not b: return 0.0
-        inter = len(a & b); union = len(a | b)
-        return inter/union
-
-class KPIFeedback:
-    """直近ウィンドウでミニKPIを計算し、隠しプロンプトと調整案を返す。"""
-    def __init__(self, cfg: MeetingConfig):
-        self.cfg = cfg
-        self._last_hint = None
-
-    def assess(self, turns: List[Turn], unresolved_hist: List[int]) -> Dict[str, typing.Any]:
-        W = max(3, int(self.cfg.kpi_window))
-        window = turns[-W:] if len(turns) >= W else turns[:]
-        if len(window) < 3:
-            return {}
-        texts = [t.content for t in window]
-        # diversity（Jaccardの連続平均を反転）
-        sims = []
-        for i in range(len(texts)-1):
-            a = self._token_set(texts[i]); b = self._token_set(texts[i+1])
-            sims.append(self._jacc(a,b))
-        diversity = 1 - (sum(sims)/len(sims) if sims else 0.0)
-        # decision density
-        decision_words = ("決定","合意","採用","実施","次回","担当","期限")
-        hits = sum(1 for t in texts if any(w in t for w in decision_words))
-        decision_density = hits/max(1,len(texts))
-        # progress（未解決の停滞を簡易判定）
-        stall = False
-        if len(unresolved_hist) >= min(4, W):
-            recent = unresolved_hist[-min(4, W):]
-            non_increasing = all(recent[i] <= recent[i-1] for i in range(1, len(recent)))
-            no_change = len(set(recent)) == 1
-            stall = no_change or (non_increasing and not any(recent[i] < recent[i-1] for i in range(1, len(recent)))) is False and \
-                (len(set(recent)) == 1)
-
-        actions: Dict[str, typing.Any] = {
-            "metrics": {
-                "diversity": round(diversity, 3),
-                "decision_density": round(decision_density, 3),
-                "stall": bool(stall)}}
-        hints = []
-        tune = {}
-        # 規則1：多様性不足 → 新観点 + 反復抑制
-        if diversity < self.cfg.th_diversity_min:
-            if self.cfg.kpi_auto_tune:
-                tune["select_temp"] = ("inc", 0.20, 0.7, 1.5)     # よりランダムに
-                tune["sim_penalty"] = ("inc", 0.10, 0.15, 0.60)   # 現在値に+0.10、最小+上限
-            else:
-                hints.append("新しい観点を必ず1つだけ追加し、直前の発言にない要素を入れてください。")
-        # 規則2：決定密度不足 → 担当+期限の強制
-        if decision_density < self.cfg.th_decision_min:
-            if self.cfg.kpi_auto_tune:
-                tune["cooldown"] = ("inc", 0.05, 0.10, 0.35)      # 連投抑制
-            else:
-                hints.append("次の発言には担当者と期限を必ず1行で含めてください（例: 担当:A、期限:9/30）。")
-        # 規則3：進捗停滞 → 収束ショック
-        if stall:
-            hints.append("抽象を避け、数値・手順・失敗時対策を各1行で具体化してください。")
-            tune["shock_mode"] = "exploit"
-
-        if not hints and not tune:
-            return actions
-        actions["hint"] = " / ".join(hints)
-        actions["tune"] = tune
-        return actions
-
-    @staticmethod
-    def _token_set(text: str) -> set:
-        t = re.sub(r"[0-9]+"," ", text); t = re.sub(r"[^\w\u3040-\u30ff\u4e00-\u9fff]+"," ", t)
-        return set(w for w in t.lower().split() if len(w)>1)
-    @staticmethod
-    def _jacc(a:set,b:set)->float:
-        if not a or not b: return 0.0
-        return len(a&b)/len(a|b)
-# ---- Step 5: ShockEngine（裏方AI：奇想天外な刺激を注入）----
-class ShockEngine:
-    """
-    mode:
-      - random  : 無方向のアイデア飛び石
-      - explore : 発散寄り（新規性を強調）
-      - exploit : 収束寄り（具体化・制約強化）
-    出力は“非公開ヒント”。本文には出さず、次の数ターンにだけ影響。
-    """
-    def __init__(self, cfg: MeetingConfig):
-        self.mode = cfg.shock
-
-    def generate(self, ctx: Dict) -> Tuple[str, str]:
-        topic = ctx.get("topic","")
-        reason = ctx.get("reason","")
-        summ = ctx.get("summary","")
-        if self.mode == "explore":
-            hint = self._hint_explore(topic, reason, summ)
-            note = "explore"
-        elif self.mode == "exploit":
-            hint = self._hint_exploit(topic, reason, summ)
-            note = "exploit"
-        else:
-            hint = self._hint_random(topic, reason, summ)
-            note = "random"
-        return hint, note
-
-    def _hint_random(self, topic, reason, summ) -> str:
-        seeds = [
-            "制約を1つ極端に強めてみて（時間=120秒、道具=1点、人数=2名など）。",
-            "逆転発想：目的を真逆にすると何が生まれる？",
-            "異分野の比喩を1つだけ移植して（料理/ダンス/将棋/交通のいずれか）。",
-            "禁止語を1つ決めて、それを回避するルールを作って。",
-        ]
-        return random.choice(seeds)
-
-    def _hint_explore(self, topic, reason, summ) -> str:
-        return "現状の発想の前提を1つ外して、全く別系統の案を1つだけ提示して。実験可能性は低くても良い。"
-
-    def _hint_exploit(self, topic, reason, summ) -> str:
-        return "直前の合意要素を3つ選び、数値・手順・失敗時対策を各1行で具体化して。抽象語は禁止。"
-
-class PendingTracker:
-    """要約から“残課題・リスク・改善案”を抽出して蓄積し、重複を除去して持つ。"""
-    KEYS = ("残課題", "課題", "リスク", "改善", "是正", "対策")
-    def __init__(self):
-        self.items = set()
-    def add_from_text(self, text: str):
-        for line in text.splitlines():
-            s = line.strip(" ・-*\t")
-            if not s:
-                continue
-            if any(k in s for k in self.KEYS):
-                # “: ”以降や「- 」以降を抽出して短文化
-                s = re.sub(r'^[^:：]*[:：]\s*', '', s)
-                self.items.add(s)
-    def clear(self):
-        self.items.clear()
 
 class Meeting:
     def __init__(self, cfg: MeetingConfig):
